+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Fixes
 
 * Upgrade some dependencies to resolve some security vulnerabilities (GUIC-208)
+* Fix the `focus` event of Input to always emit the current value of the input field (GUIC-160).
+* Fix the handling of the `DateTimePicker.timestamp` input property - it can now be used to set a new value and not just the initial value (GUIC-160).
 
 ## 6.6.0 (2018-02-05)